update tests, add new one with isset
