+++ options-gen/testdata/case-01-empty-options/options_generated.go
@@ -2,21 +2,12 @@
 
 package testcase
 
-type optField int8
-
-const ()
-
-var optIsSet = [0]bool{}
-
 type OptOptionsSetter func(o *Options)
 
 func NewOptions(
 	options ...OptOptionsSetter,
 ) Options {
 	o := Options{}
-
-	var empty [0]bool
-	optIsSet = empty
 
 	// Setting defaults from field tag (if present)
 
@@ -28,8 +19,4 @@ func NewOptions(
 
 func (o *Options) Validate() error {
 	return nil
-}
-
-func (o *Options) IsSet(field optField) bool {
-	return optIsSet[field]
 }
