+++ options-gen/testdata/case-16-without-cosntructor/options_generated.go
@@ -2,27 +2,15 @@
 
 package testcase
 
-type optField int8
-
-const (
-	Fieldname optField = 0
-)
-
-var optIsSet = [1]bool{}
-
 type OptOptionsSetter func(o *Options)
 
 func WithName(opt string) OptOptionsSetter {
 	return func(o *Options) {
 		o.name = opt
-		optIsSet[Fieldname] = true
+
 	}
 }
 
 func (o *Options) Validate() error {
 	return nil
-}
-
-func (o *Options) IsSet(field optField) bool {
-	return optIsSet[field]
 }
