+++ options-gen/testdata/case-17-skip/options_generated.go
@@ -2,23 +2,12 @@
 
 package testcase
 
-type optField int8
-
-const (
-	Fieldname optField = 0
-)
-
-var optIsSet = [1]bool{}
-
 type OptOptionsSetter func(o *Options)
 
 func NewOptions(
 	options ...OptOptionsSetter,
 ) Options {
 	o := Options{}
-
-	var empty [1]bool
-	optIsSet = empty
 
 	// Setting defaults from field tag (if present)
 
@@ -31,14 +20,10 @@ func NewOptions(
 func WithName(opt string) OptOptionsSetter {
 	return func(o *Options) {
 		o.name = opt
-		optIsSet[Fieldname] = true
+
 	}
 }
 
 func (o *Options) Validate() error {
 	return nil
-}
-
-func (o *Options) IsSet(field optField) bool {
-	return optIsSet[field]
 }
