+++ options-gen/testdata/case-20.2-isset-empty-options/.params.json
@@ -1,0 +1,3 @@
+{
+  "with_isset": true
+}
