+++ options-gen/testdata/case-20.2-isset-empty-options/options.go
@@ -1,0 +1,3 @@
+package testcase
+
+type Options struct{}
